Significant rewrite of game logic.
In dealing with the game in another format, I realized that the
Cell class was completely superfluous, and the game could be done
without it.  I rewrote all the logic inside the Grid class to no
longer utilize Cell, and then deleted the Cell class altogether.

The demo file was also updated to use 1 and 0 for live and dead
cells rather than letter strings.